Refactor upe folders, classic and styles
