+++ webpack.config.js
@@ -12,6 +12,6 @@ module.exports = {
 	],
 	entry: {
 		index: './client/blocks/index.js',
-		upe_classic: './client/classic/upe.js',
+		upe_classic: './client/classic/upe/index.js',
 	}
 };
